Refactor helper
